Update the Download page to use `cs setup`.
Also recommend VS Code with Metals.

Code-wise, we inline the data that were in `_data/downloads.yml`
and `_data/tutorials.yml` as well as inline and unroll the logic
that was in `_includes/tutorial-list.html`. Those were only used
once, and it is much easier to edit them in-place, not to mention
that it is actually works with jekyll's incremental builds, unlike
the `_data`-based approach.